# internal/repositories/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Budget struct {
	ID          uuid.UUID          `json:"id"`
	WorkspaceID uuid.UUID          `json:"workspace_id"`
	Year        int16              `json:"year"`
	Month       int16              `json:"month"`
	TotalLimit  float64            `json:"total_limit"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type BudgetCategory struct {
	BudgetID    uuid.UUID `json:"budget_id"`
	CategoryID  uuid.UUID `json:"category_id"`
	LimitAmount float64   `json:"limit_amount"`
}

type Category struct {
	ID          uuid.UUID          `json:"id"`
	WorkspaceID *uuid.UUID         `json:"workspace_id"`
	Name        string             `json:"name"`
	Icon        pgtype.Text        `json:"icon"`
	Color       pgtype.Text        `json:"color"`
	Type        string             `json:"type"`
	IsSystem    bool               `json:"is_system"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Debt struct {
	ID               uuid.UUID          `json:"id"`
	WorkspaceID      uuid.UUID          `json:"workspace_id"`
	Name             string             `json:"name"`
	Lender           pgtype.Text        `json:"lender"`
	Principal        float64            `json:"principal"`
	Rate             float64            `json:"rate"`
	RateType         string             `json:"rate_type"`
	Installments     int32              `json:"installments"`
	FirstPaymentDate pgtype.Date        `json:"first_payment_date"`
	Notes            pgtype.Text        `json:"notes"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
}

type DebtPayment struct {
	ID        uuid.UUID          `json:"id"`
	DebtID    uuid.UUID          `json:"debt_id"`
	Period    int32              `json:"period"`
	Amount    float64            `json:"amount"`
	PaidAt    pgtype.Date        `json:"paid_at"`
	Notes     pgtype.Text        `json:"notes"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type SavingsContribution struct {
	ID            uuid.UUID          `json:"id"`
	GoalID        uuid.UUID          `json:"goal_id"`
	Amount        float64            `json:"amount"`
	ContributedAt pgtype.Date        `json:"contributed_at"`
	Notes         pgtype.Text        `json:"notes"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
}

type SavingsGoal struct {
	ID           uuid.UUID          `json:"id"`
	WorkspaceID  uuid.UUID          `json:"workspace_id"`
	Name         string             `json:"name"`
	TargetAmount float64            `json:"target_amount"`
	Deadline     pgtype.Date        `json:"deadline"`
	Notes        pgtype.Text        `json:"notes"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type Transaction struct {
	ID                uuid.UUID          `json:"id"`
	WorkspaceID       uuid.UUID          `json:"workspace_id"`
	UserID            uuid.UUID          `json:"user_id"`
	CategoryID        *uuid.UUID         `json:"category_id"`
	TransferID        *uuid.UUID         `json:"transfer_id"`
	Type              string             `json:"type"`
	TransferDirection pgtype.Text        `json:"transfer_direction"`
	Amount            float64            `json:"amount"`
	Description       pgtype.Text        `json:"description"`
	Date              pgtype.Date        `json:"date"`
	CreatedAt         pgtype.Timestamptz `json:"created_at"`
	UpdatedAt         pgtype.Timestamptz `json:"updated_at"`
}

type Transfer struct {
	ID              uuid.UUID          `json:"id"`
	FromWorkspaceID uuid.UUID          `json:"from_workspace_id"`
	ToWorkspaceID   uuid.UUID          `json:"to_workspace_id"`
	Note            pgtype.Text        `json:"note"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID           uuid.UUID          `json:"id"`
	Email        string             `json:"email"`
	PasswordHash string             `json:"password_hash"`
	Name         string             `json:"name"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type Workspace struct {
	ID        uuid.UUID          `json:"id"`
	Name      string             `json:"name"`
	OwnerID   uuid.UUID          `json:"owner_id"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	Currency  string             `json:"currency"`
}

type WorkspaceInvitation struct {
	ID          uuid.UUID          `json:"id"`
	WorkspaceID uuid.UUID          `json:"workspace_id"`
	Email       string             `json:"email"`
	Role        string             `json:"role"`
	Token       uuid.UUID          `json:"token"`
	Status      string             `json:"status"`
	InvitedBy   uuid.UUID          `json:"invited_by"`
	ExpiresAt   pgtype.Timestamptz `json:"expires_at"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type WorkspaceMember struct {
	WorkspaceID uuid.UUID          `json:"workspace_id"`
	UserID      uuid.UUID          `json:"user_id"`
	Role        string             `json:"role"`
	JoinedAt    pgtype.Timestamptz `json:"joined_at"`
}